requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2
SQLAlchemy>=2.0
psycopg[binary]==3.2.1
psycopg2-binary
openai==1.46.0
boto3
botocore
google-auth>=2.23.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
device-info>=0.1.0
alembic